Genericise conditions
