+++ Assets/ExpandedStationA/ExpandStationScene.unity
@@ -66110,6 +66110,11 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 970000952001712002, guid: e622483ac854b05458d1f78890cf5230,
+        type: 3}
+      propertyPath: conditions.Array.size
+      value: 1
+      objectReference: {fileID: 0}
     - target: {fileID: 1617741008469135249, guid: e622483ac854b05458d1f78890cf5230,
         type: 3}
       propertyPath: m_RootOrder
