+++ Assets/Scripts/Animation/Timeline/CutsceneCollisionTrigger.cs
@@ -4,30 +4,14 @@ using UnityEngine;
 public class CutsceneCollisionTrigger : MonoBehaviour
 {
     [SerializeField] private Cutscene cutscene;
-    [Header("Leave blank if not used in the condition")]
-    [SerializeField] private string dialogueBoolVariableName;
-    [SerializeField] private Mission mission;
+    [Header("Leave all blank that aren't used in the condition")]
+    [SerializeField] private Condition[] conditions;
 
     private BoxCollider boxCollider;
 
     protected virtual bool CutsceneTriggerable(Collider other)
     {
-        if (!string.IsNullOrWhiteSpace(dialogueBoolVariableName))
-        {
-            var value = DialogueDatabaseManager.GetLuaVariableAsBool(dialogueBoolVariableName);
-            Debug.Log("CutsceneCollisionTrigger dialogue variable value: " + value);
-            
-            if (!value)
-            {
-                return false;
-            }
-        }
-
-        if (mission != null && !mission.HasBeenCompleted)
-        {
-            return false;
-        }
-        return other.CompareTag(PlayerConstants.PlayerTag);
+        return other.CompareTag(PlayerConstants.PlayerTag) && conditions.AreAllMet();
     }
 
     private void Awake()
