+++ Assets/Scripts/Common.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 95ffdaf7e2876494e9a59552ee8df544
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
