+++ Assets/Scripts/Common/Condition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c0c131e98f625d844812d7c3176de292
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
